builds/tests/requirements.txt
selenium
pytest
pytest-xdist
requests
allure-pytest
Faker
sqlalchemy
pymysql